+++ tools/eslint-plugin-ng-elemental/index.mjs
@@ -1,0 +1,9 @@
+import bemHostClass from './rules/bem-host-class.mjs';
+import bemTemplateClass from './rules/bem-template-class.mjs';
+
+export default {
+  rules: {
+    'bem-host-class': bemHostClass,
+    'bem-template-class': bemTemplateClass,
+  },
+};